Update README.md
Write a program for motion estimation (ME) using the block matching methods on the given video sequence. You have to implement two search algorithms to find motion vectors, the full search and the 2D logarithmic search method. The search range is ± pixels along horizontal and vertical directions. In this implementation, you should apply the ME to all non-overlapping macroblocks to evaluate the motion vectors, and the block matching measure is defined as sum of absolute differences (SAD), which is described in the slide (p.48).